requirements.txt
pydantic-ai==0.0.8
gradio==5.7.1
groq==0.13.0
yfinance==0.2.50
python-dotenv==1.0.1
